refactor(goal): split terminal frontier read model
