+++ loopx/canary/quality_surface_catalog.py
@@ -181,7 +181,8 @@ QUALITY_SURFACE_CATALOG: tuple[dict[str, Any], ...] = (
         "risk": "high",
         "canary_profile_id": "goal-frontier-replan-rules",
         "owner_paths": [
-            "loopx/control_plane/goals/goal_frontier.py",
+            "loopx/control_plane/goals/goal_frontier/__init__.py",
+            "loopx/control_plane/goals/goal_frontier/terminal.py",
             "loopx/control_plane/goals/goal_frontier_replan_rules.py",
             "loopx/control_plane/work_items/autonomous_replan_ack.py",
         ],
